+++ src/manager.ts
@@ -1,0 +1,1 @@
+export * from 'storybook-i18n/manager';