+++ src/preset/preview.ts
@@ -1,3 +1,0 @@
-import {withI18Next} from '../withI18Next';
-
-export const decorators = [withI18Next];